Switch the unbounded dropwizard metrics queues to bounded queues
If the metrics2 subsystem isn't up, we want to make sure we don't
hold an excessive number of metrics in memory.

* Moves debug logs to trace (not really "debug" appropriate) and fix some javadoc warnings.
* Address sidseth's suggestion about metrics to add which exceed the maximum limit of the queue
* Add in a standalone utility which sets up a FileSink and generates some metrics to observe the impact of the changes.

Additionally, this adds synchronize(this) during the metrics2 collection phase

ScheduledReporter is doing a synchronized(this) before calling
report() on our Reporter implementation. This will make sure there
aren't any unintended side-effects if they are both called at the beginning.

Closes #2, Closes #3