Ensure that the dropwizard elements are always reset after iteration
In the case which we error while processing some dropwizard metrics into
a metrics2 builder, we do not want continually reattempt to process the
same records over and over again. If we fail to parse all of the metrics,
we should still mark them as done (remove them).

Minor issue since the next cycle of the reporter would reset things anyways
but this is a bit more clean

Closes #5